Rename component and delete styles props
