+++ src/components/molecules/PostMolecule.tsx
@@ -1,5 +1,5 @@
 import { RouterOutputs } from "~/utils/api";
-import Avatar from "./Avatar";
+import AvatarAtom from "../atoms/AvatarAtom";
 
 // 上記のtypeと同じですが、APIから取得してるデータで作成みたい。
 type Post = RouterOutputs["posts"]["index"][number];
@@ -7,10 +7,9 @@ type Post = RouterOutputs["posts"]["index"][number];
 const PostMolecule = ({post, author}: Post) => {
   return (
     <div key={post.id}className="p-8 border-b gap-4 border-slate-700 flex items-center">
-      <Avatar 
+      <AvatarAtom
         src={author.imageUrl} 
         alt="User profile avatar" 
-        styles="rounded-full"
         width={40}
         height={40}
       />
